+++ src/main/java/roomescape/infrastructure/persistence/JdbcThemeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.infrastructure;
+package roomescape.infrastructure.persistence;
 
 import java.util.List;
 import java.util.Objects;
@@ -10,7 +10,6 @@ import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.domain.Theme;
 import roomescape.domain.ThemeRepository;
-import roomescape.infrastructure.rowmapper.ThemeRowMapper;
 
 @Repository
 public class JdbcThemeRepository implements ThemeRepository {
