+++ src/test/java/roomescape/infrastructure/persistence/JdbcReservationTest.java
@@ -1,4 +1,4 @@
-package roomescape.infrastructure;
+package roomescape.infrastructure.persistence;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
