+++ src/test/java/roomescape/infrastructure/persistence/JdbcThemeRepositoryTest.java
@@ -1,4 +1,4 @@
-package roomescape.infrastructure;
+package roomescape.infrastructure.persistence;
 
 import static org.assertj.core.api.Assertions.assertThat;
